backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
google-cloud-aiplatform>=1.39.0
google-auth>=2.26.1
httpx==0.26.0
langchain>=0.1.0
langgraph>=0.0.15
langchain-google-vertexai>=0.0.3
asyncpg>=0.29.0
pgvector>=0.2.4
cloud-sql-python-connector[asyncpg]>=1.4.0
google-cloud-discoveryengine>=0.11.0
langchain-google-genai>=0.0.3
google-cloud-speech>=2.26.0

